requirements.txt
plotly
pandas
python-telegram-bot